compiling, even some tests ok

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: gostroum <gostroum@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/27 01:43:44 by gostroum          #+#    #+#             */
-/*   Updated: 2025/05/27 19:43:59 by gostroum         ###   ########.fr       */
+/*   Updated: 2025/05/27 21:02:36 by gostroum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct s_stash
 {
 	char	*s;
 	int		fd;
-	int		len;
+	ssize_t	len;
 }	t_stash;
 
 char	*get_next_line(int fd);
